名前変更
オブジェクトの名前を一部変更

+++ Christmas-Gift/Project/GameProject/CollisionHitBox/CollisionBoxBase.h
@@ -8,7 +8,7 @@ class CollisionBoxBase : public Task
 {
 public:
 	ToolTips* m_tooltips;
-	int m_time;
+	float m_time;
 	//コンストラクタ
 	CollisionBoxBase(ETaskTag tag,bool remove);
